--- src/routes/CreateRoom.js
@@ -1,15 +1,0 @@
-import React from "react";
-import { v1 as uuid } from "uuid";
-
-const CreateRoom = (props) => {
-    function create() {
-        const id = uuid();
-        props.history.push(`/room/${id}`);
-    }
-
-    return (
-        <button onClick={create}>Create room</button>
-    );
-};
-
-export default CreateRoom;
